Fix bug with rx gain

+++ radioop.h
@@ -155,7 +155,7 @@ private:
     long long _tune_shift_freq;
     float _tx_power;
     int _squelch;
-    int _rx_sensitivity;
+    float _rx_sensitivity;
     int _step_hz;
     int _tune_limit_lower;
     int _tune_limit_upper;
